warden: Disable AppArmor if possible
If AppArmor is configured on the host, it may interfere with processes
running inside containers, as it isn't aware that these processes have a
different root file system. For instance: files these processes try to
access appear to AppArmor as relative to the hosts root file system, and
may by policy be therefore denied.

Change-Id: I70d5acfccebdd1d4cc4b6e86a806b34e503f369d