Relax iptables rules for NAT
Traffic from one container to another on the same machine should be
allowed as long as the originating container sends traffic to one of the
publicly mapped ports of the other container.

To allow this the PREROUTING chain must work on traffic received on any
interface, and the POSTROUTING chain should masquerade traffic sent on
every interface.

Change-Id: I8255e7847ebbac6eb02e21f0fb34a2f7d3de6203