support for running linux backend remotely
This changes all usage of exec.Command to manually constructing, so that it
doesn't resolve the executable path locally. A RemoteCommandRunner is added
that wraps the command with SSH.

Added a private network to the Vagrantfile and removed the shared folders for
running Go. The Vagrant VM should now be able to just run in the background
and the Garden server itself can be run locally.