fix --continue flag remaining after setting title
This is actually pretty hacky. There's no clean way to set the process title,
so this passes argv[0] as whitespace with the size of the requested title. We
then detect whether to daemonize based on whether argv[0] is whitespace
instead of using a flag. This is to avoid having to clear out argv[1].

The exec'd process then fills in the whitespace with the requested title.