add performance test exercising streaming API
This test is to exercise the Stream API with a job that's constantly churning
out a lot of output. A loop that just keeps echoing to stdout and stderr. It
starts N streams (where N is from 0 to 100 in increments of 10), and exercises
the rest of the API, measuring the time they took, and the throughput of all
of the streams.