remove Go wshd/wsh
This was written to further my understanding of the lower-level bits. In
practice, writing such low-level code in Go actually becomes more convoluted
than it would be with C, with few benefits.

The one nice bit that stayed as pretty high-level Go was the 'daemon' package,
which served requests for running commands. However, to support setting
rlimits, this would have to get a bit uglier than just using os/exec. So it's
probably time to just cut our losses.

I've kept the tests and moved them under integration/ - as they're still nice
to have and should "just work" with C wshd.

Also, in running the stream integration tests, it shows that the C wshs is
quite a bit faster, and probably has higher throughput.