command runner: don't propagate signals to child
These commands are expected to be unaffected by the server receiving signals,
as the server will intercept TERM/INT to perform a clean shutdown and stop
handling requests. Most requests are handled using the command runner, so
forwarding the signals is counter productive, as the commands are expected to
complete normally.

[#61232150]