reap containers that are idle beyond grace time
The server can be configured with a global default grace time via a flag,
otherwise the value given in the Create message is respected.

Note: the grace time given to a container on creation, whether it's via the
message or the server default, is persisted in the snapshot. This means that
if the default value is changed, previous containers are *not* switched to the
new default.

[finishes #61232348]