detect flat cgroup mount, mount cgroup accordingly
If an existing cgroup mount is mounted flat, all other cgroup mounts have to
be flat as well. Similarly, if there are existing mounts for individual
subsystems, other mounts have to be done the same. So, detect if there's an
existing flat mount, and mount ours flat as well, and just bind-mount each
subsystem.