simply umount the pivoted root with MNT_DETACH
no longer recurse through the sub-mount points; this is prone to error if e.g.
a mount point's source is deleted. in this state, some kernels will not allow
it to be umounted.

this apprach is much simpler and seems to be just as safe (none of the mount
points are visible in the child).