Avoid race condition when process exits while streaming stdin
When a process exits the server stops listening, and so sending stdin will fail.
Previously when sending stdin failed the client
closed the connection to the server, which could race with reading
the final process exit status. This would cause a .Wait() on the process
to sometimes exit with an error rather than an exit code even though the
process had completely succesfully.

Signed-off-by: Craig Furman <cfurman@pivotal.io>